backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Hashimoto Joshua <coffeesteam.jh38+git@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "0.112.2"
python-jose = "3.3.0"
passlib = "1.7.4"
pydantic-settings = "2.1.0"
asyncpg = "0.29.0"
ormar = "^0.20.1"
psycopg2-binary = "^2.9.9"


[tool.poetry.group.dev.dependencies]
ruff = "^0.6.2"


[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "migrations",
]
select = [
    "F",
    "E",
    "W",
    "I",
    "B",
    "PL",
    "PD",
    "NPY",
    "DJ",
    "RUF",
    "UP",
]
ignore = [
    "B008",
]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
